+++ src/main/java/com/hello/background/vo/SalarySpecialItemVO.java
@@ -41,7 +41,7 @@ public class SalarySpecialItemVO implements Serializable {
     /**
      * 金额
      */
-    private Integer sum;
+    private Double sum;
     /**
      * 地点
      */
